+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Duplication of upgrade parts in some cases [#121]
+
 ## [1.6.4] - 2020-05-13
 ### Added
 - Indestructible trait, prevents durability loss (not found on any default part for now, but the trait does work)